% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{first_last}
\alias{first_last}
\title{First and last non-NA of a vector}
\usage{
first_last(v)
}
\arguments{
\item{v}{numeric. @return length-2 vector c(first, last). @export}
}
\description{
First and last non-NA of a vector
}
\examples{
first_last(c(NA, 3, 5, NA, 8, NA))
}
