% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{interpret_nap}
\alias{interpret_nap}
\title{Interpret a NAP value (Parker & Vannest 2009 bands)}
\usage{
interpret_nap(x)
}
\arguments{
\item{x}{a NAP value. @return a Japanese effect-size label. @export}
}
\description{
Interpret a NAP value (Parker & Vannest 2009 bands)
}
\examples{
interpret_nap(0.95)
interpret_nap(0.70)
}
